Restructure data files folders
