+++ database/insertData.js
@@ -4,16 +4,16 @@ const fetch = require('node-fetch')
 const config = require('../src/config.json')
 
 const dataFolder = './database/insertData'
-const sharedDataFiles = ['core_mutations.js', 'dev_actions.js']
+// const sharedDataFiles = ['core_mutations.js', 'dev_actions.js']
 
 if (process.argv[2] === '--from_insert_data.sh') {
   const filesToProcess = fs
-    .readdirSync(dataFolder, { withFileTypes: true })
+    .readdirSync(path.join(dataFolder, '_common'), { withFileTypes: true })
     .filter((dirent) => dirent.isFile())
-    .map((dirent) => dirent.name)
+    .map((dirent) => path.join('_common', dirent.name))
     .filter((file) => !file.match(/^\./)) // Ignore hidden files
-    .filter((file) => !sharedDataFiles.includes(file))
 
+  console.log('filesToProcess', filesToProcess)
   // Add locale-specific files
   const localeFolder = process.argv[3] || 'dev'
   if (localeFolder) {
@@ -21,13 +21,8 @@ if (process.argv[2] === '--from_insert_data.sh') {
     const subfolderFilesToProcess = fs
       .readdirSync(path.join(dataFolder, localeFolder))
       .filter((file) => !file.match(/^\./)) // Ignore hidden files
-      .filter((file) => !sharedDataFiles.includes(file))
     filesToProcess.push(
       ...subfolderFilesToProcess.map((filename) => path.join(localeFolder, filename))
-    )
-    // Copy current version of core mutations/actions to subfolder
-    sharedDataFiles.forEach((file) =>
-      fs.copyFileSync(path.join(dataFolder, file), path.join(dataFolder, localeFolder, file))
     )
   }
