+++ database/insertData/_common/02D_template_categories.js
