+++ database/insertData/laos/03_template_B_orgRegistration.js
@@ -8,8 +8,8 @@ TEMPLATE B - Organisation Registration
   - for creating a new Organisation in the system. Requires single review
   by reviewer with "reviewOrgRego" permission
 */
-const { coreActions, joinFilters } = require('./core_mutations')
-const { devActions } = require('./dev_actions')
+const { coreActions, joinFilters } = require('../_helpers/core_mutations')
+const { devActions } = require('../_helpers/dev_actions')
 
 exports.queries = [
   `mutation {
